finder: split aggregation logic from finder logic
this modularization will unlock better testing in a next patch.
