+++ pkg/finder/finder.go
@@ -9,7 +9,6 @@ import (
 	"log"
 	"net"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"time"
 
@@ -18,19 +17,13 @@ import (
 	"google.golang.org/grpc"
 
 	"k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	pb "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
 	"k8s.io/kubernetes/pkg/kubelet/cm/cpuset"
 	"k8s.io/kubernetes/pkg/kubelet/util"
-
-	"github.com/fromanirh/numalign/pkg/topologyinfo/cpus"
-	"github.com/fromanirh/numalign/pkg/topologyinfo/pcidev"
-	v1alpha1 "github.com/swatisehgal/topologyapi/pkg/apis/topology/v1alpha1"
 )
 
 const (
 	defaultTimeout     = 5 * time.Second
-	ns                 = "resource-topology-exporter"
 	CGroupCPUSetPrefix = "fs/cgroup/cpuset"
 	CGroupCPUsetSuffix = "cpuset.cpus"
 )
@@ -47,22 +40,16 @@ type Args struct {
 
 type CRIFinder interface {
 	Scan() ([]PodResources, error)
-	Aggregate(podResData []PodResources) []v1alpha1.NUMANodeResource
 }
 
 type CGroupPathTranslator func(sysfs, cgroupPath string) string
 
 type criFinder struct {
+	// we may want to move to cadvisor past PoC stage
 	args            Args
 	conn            *grpc.ClientConn
 	client          pb.RuntimeServiceClient
 	cgroupTranslate CGroupPathTranslator
-	// we may want to move to cadvisor past PoC stage
-	pciDevs         *pcidev.PCIDevices
-	cpus            *cpus.CPUs
-	cpuID2NUMAID    map[int]int
-	pciAddr2NUMAID  map[string]int
-	perNUMACapacity map[int]map[v1.ResourceName]int64
 	// pciaddr -> resourcename
 	pci2ResourceMap map[string]string
 }
@@ -79,46 +66,11 @@ type ContainerInfo struct {
 	RuntimeSpec    *runtimespec.Spec   `json:"runtimeSpec"`
 }
 
-func makePCI2ResourceMap(numaNodes int, pciDevs *pcidev.PCIDevices, pciResMapConf map[string]string) (map[int]map[v1.ResourceName]int64, map[string]string, map[string]int) {
-	pciAddr2NUMAID := make(map[string]int)
-	pci2Res := make(map[string]string)
-
-	perNUMACapacity := make(map[int]map[v1.ResourceName]int64)
-	for nodeNum := 0; nodeNum < numaNodes; nodeNum++ {
-		perNUMACapacity[nodeNum] = make(map[v1.ResourceName]int64)
-
-		for _, pciDev := range pciDevs.Items {
-			if pciDev.NUMANode() != nodeNum {
-				continue
-			}
-			sriovDev, ok := pciDev.(pcidev.SRIOVDeviceInfo)
-			if !ok {
-				continue
-			}
-
-			if !sriovDev.IsVFn {
-				continue
-			}
-
-			resName, ok := pciResMapConf[sriovDev.ParentFn]
-			if !ok {
-				continue
-			}
-
-			pci2Res[sriovDev.Address()] = resName
-			pciAddr2NUMAID[sriovDev.Address()] = nodeNum
-			perNUMACapacity[nodeNum][v1.ResourceName(resName)]++
-		}
-	}
-	return perNUMACapacity, pci2Res, pciAddr2NUMAID
-}
-
-func NewFinder(args Args, pciResMapConf map[string]string) (CRIFinder, error) {
+func NewFinder(args Args, pci2ResourceMap map[string]string) (CRIFinder, error) {
 	finderInstance := &criFinder{
 		args:            args,
-		perNUMACapacity: make(map[int]map[v1.ResourceName]int64),
+		pci2ResourceMap: pci2ResourceMap,
 	}
-	var err error
 
 	// At this stage, we only support containerd and cri-o
 	if args.ContainerRuntime == "containerd" {
@@ -126,48 +78,6 @@ func NewFinder(args Args, pciResMapConf map[string]string) (CRIFinder, error) {
 	} else {
 		//cri-o
 		finderInstance.cgroupTranslate = crioCGroupPathTranslate
-	}
-
-	// first scan the sysfs
-	// CAUTION: these resources are expected to change rarely - if ever. So we are intentionally do this once during the process lifecycle.
-	finderInstance.cpus, err = cpus.NewCPUs(finderInstance.args.SysfsRoot)
-	if err != nil {
-		return nil, fmt.Errorf("error scanning the system CPUs: %v", err)
-	}
-	for nodeNum, cpuList := range finderInstance.cpus.NUMANodeCPUs {
-		log.Printf("detected system CPU: NUMA cell %d cpus = %v\n", nodeNum, cpuList)
-	}
-
-	for nodeNum := 0; nodeNum < finderInstance.cpus.NUMANodes; nodeNum++ {
-	}
-
-	finderInstance.pciDevs, err = pcidev.NewPCIDevices(finderInstance.args.SysfsRoot)
-	if err != nil {
-		return nil, fmt.Errorf("error scanning the system PCI devices: %v", err)
-	}
-	for _, pciDev := range finderInstance.pciDevs.Items {
-		log.Printf("detected system PCI device = %s\n", pciDev.String())
-	}
-
-	// helper maps
-	var pciDevMap map[int]map[v1.ResourceName]int64
-	pciDevMap, finderInstance.pci2ResourceMap, finderInstance.pciAddr2NUMAID = makePCI2ResourceMap(finderInstance.cpus.NUMANodes, finderInstance.pciDevs, pciResMapConf)
-	finderInstance.cpuID2NUMAID = make(map[int]int)
-	for nodeNum, cpus := range finderInstance.cpus.NUMANodeCPUs {
-		for _, cpu := range cpus {
-			finderInstance.cpuID2NUMAID[cpu] = nodeNum
-		}
-	}
-
-	// initialize with the capacities
-	for nodeNum := 0; nodeNum < finderInstance.cpus.NUMANodes; nodeNum++ {
-		finderInstance.perNUMACapacity[nodeNum] = make(map[v1.ResourceName]int64)
-		for resName, count := range pciDevMap[nodeNum] {
-			finderInstance.perNUMACapacity[nodeNum][resName] = count
-		}
-
-		cpus := finderInstance.cpus.NUMANodeCPUs[nodeNum]
-		finderInstance.perNUMACapacity[nodeNum][v1.ResourceCPU] = int64(len(cpus))
 	}
 
 	// now we can connext to CRI
@@ -196,7 +106,7 @@ func getAddressAndDialer(endpoint string) (string, func(ctx context.Context, add
 	return util.GetAddressAndDialer(endpoint)
 }
 
-func (f *criFinder) listContainersResponse() (*pb.ListContainersResponse, error) {
+func (cpf *criFinder) listContainersResponse() (*pb.ListContainersResponse, error) {
 	st := &pb.ContainerStateValue{}
 	st.State = pb.ContainerState_CONTAINER_RUNNING
 	filter := &pb.ContainerFilter{}
@@ -206,7 +116,7 @@ func (f *criFinder) listContainersResponse() (*pb.ListContainersResponse, error)
 		Filter: filter,
 	}
 
-	ListContResponse, err := f.client.ListContainers(context.Background(), ListContReq)
+	ListContResponse, err := cpf.client.ListContainers(context.Background(), ListContReq)
 	if err != nil {
 		fmt.Errorf("Error in  ListContResponse: %v", err)
 		return nil, err
@@ -214,12 +124,12 @@ func (f *criFinder) listContainersResponse() (*pb.ListContainersResponse, error)
 	return ListContResponse, nil
 }
 
-func (f *criFinder) containerStatsResponse(c *pb.Container) (*pb.ContainerStatsResponse, error) {
+func (cpf *criFinder) containerStatsResponse(c *pb.Container) (*pb.ContainerStatsResponse, error) {
 	//ContainerStats
 	ContStatsReq := &pb.ContainerStatsRequest{
 		ContainerId: c.Id,
 	}
-	ContStatsResp, err := f.client.ContainerStats(context.Background(), ContStatsReq)
+	ContStatsResp, err := cpf.client.ContainerStats(context.Background(), ContStatsReq)
 	if err != nil {
 		log.Printf("Error in  ContStatsResp: %v", err)
 		return nil, err
@@ -227,13 +137,13 @@ func (f *criFinder) containerStatsResponse(c *pb.Container) (*pb.ContainerStatsR
 	return ContStatsResp, nil
 }
 
-func (f *criFinder) containerStatusResponse(c *pb.Container) (*pb.ContainerStatusResponse, error) {
+func (cpf *criFinder) containerStatusResponse(c *pb.Container) (*pb.ContainerStatusResponse, error) {
 	//ContainerStatus
 	ContStatusReq := &pb.ContainerStatusRequest{
 		ContainerId: c.Id,
 		Verbose:     true,
 	}
-	ContStatusResp, err := f.client.ContainerStatus(context.Background(), ContStatusReq)
+	ContStatusResp, err := cpf.client.ContainerStatus(context.Background(), ContStatusReq)
 	if err != nil {
 		log.Printf("Error in  ContStatusResp: %v", err)
 		return nil, err
@@ -257,33 +167,6 @@ type PodResources struct {
 	Containers []ContainerResources
 }
 
-func (cpf *criFinder) updateNUMAMap(numaData map[int]map[v1.ResourceName]int64, ri ResourceInfo) {
-	if ri.Name == v1.ResourceCPU {
-		for _, cpuIDStr := range ri.Data {
-			cpuID, err := strconv.Atoi(cpuIDStr)
-			if err != nil {
-				// TODO: log
-				continue
-			}
-			nodeNum, ok := cpf.cpuID2NUMAID[cpuID]
-			if !ok {
-				// TODO: log
-				continue
-			}
-			numaData[nodeNum][ri.Name]--
-		}
-		return
-	}
-	for _, pciAddr := range ri.Data {
-		nodeNum, ok := cpf.pciAddr2NUMAID[pciAddr]
-		if !ok {
-			// TODO: log
-			continue
-		}
-		numaData[nodeNum][ri.Name]--
-	}
-}
-
 func (cpf *criFinder) listPodSandBoxResponse() (*pb.ListPodSandboxResponse, error) {
 	//ListPodSandbox
 	podState := &pb.PodSandboxStateValue{}
@@ -301,29 +184,29 @@ func (cpf *criFinder) listPodSandBoxResponse() (*pb.ListPodSandboxResponse, erro
 	return PodSbResponse, nil
 }
 
-func (f *criFinder) isWatchable(podSb *pb.PodSandbox) bool {
-	if f.args.Namespace == "" {
+func (cpf *criFinder) isWatchable(podSb *pb.PodSandbox) bool {
+	if cpf.args.Namespace == "" {
 		return true
 	}
 	//TODO:  add an explicit check for guaranteed pods
-	return f.args.Namespace == podSb.Metadata.Namespace
+	return cpf.args.Namespace == podSb.Metadata.Namespace
 }
 
-func (f *criFinder) Scan() ([]PodResources, error) {
+func (cpf *criFinder) Scan() ([]PodResources, error) {
 	//PodSandboxStatus
-	podSbResponse, err := f.listPodSandBoxResponse()
+	podSbResponse, err := cpf.listPodSandBoxResponse()
 	if err != nil {
 		return nil, err
 	}
 	var podResData []PodResources
 	for _, podSb := range podSbResponse.GetItems() {
-		if !f.isWatchable(podSb) {
+		if !cpf.isWatchable(podSb) {
 			log.Printf("SKIP pod %q\n", podSb.Metadata.Name)
 			continue
 		}
 
 		log.Printf("querying pod %q\n", podSb.Metadata.Name)
-		ListContResponse, err := f.listContainersResponse()
+		ListContResponse, err := cpf.listContainersResponse()
 		if err != nil {
 			log.Printf("fail to list containers for pod %q: err: %v", podSb.Metadata.Name, err)
 			continue
@@ -340,7 +223,7 @@ func (f *criFinder) Scan() ([]PodResources, error) {
 
 			log.Printf("querying pod %q container %q\n", podSb.Metadata.Name, c.Metadata.Name)
 
-			ContStatusResp, err := f.containerStatusResponse(c)
+			ContStatusResp, err := cpf.containerStatusResponse(c)
 			if err != nil {
 				return nil, err
 			}
@@ -371,7 +254,7 @@ func (f *criFinder) Scan() ([]PodResources, error) {
 				continue
 			}
 
-			cpus, err := getAllocatedCPUs(f.cgroupTranslate(f.args.SysfsRoot, ci.RuntimeSpec.Linux.CgroupsPath))
+			cpus, err := getAllocatedCPUs(cpf.cgroupTranslate(cpf.args.SysfsRoot, ci.RuntimeSpec.Linux.CgroupsPath))
 			if err != nil {
 				log.Printf("pod %q container %q unable to get allocatedCPUs %v as", podSb.Metadata.Name, ContStatusResp.Status.Metadata.Name, err)
 				continue
@@ -383,7 +266,7 @@ func (f *criFinder) Scan() ([]PodResources, error) {
 			}
 			contRes.Resources = append(contRes.Resources, makeCPUResource(cpuList)...)
 			if ci.Config != nil && ci.Config.Envs != nil {
-				contRes.Resources = append(contRes.Resources, makePCIDeviceResource(env, f.pci2ResourceMap)...)
+				contRes.Resources = append(contRes.Resources, makePCIDeviceResource(env, cpf.pci2ResourceMap)...)
 			}
 
 			log.Printf("pod %q container %q contData=%s\n", podSb.Metadata.Name, ContStatusResp.Status.Metadata.Name, spew.Sdump(contRes))
@@ -393,38 +276,6 @@ func (f *criFinder) Scan() ([]PodResources, error) {
 		podResData = append(podResData, podRes)
 	}
 	return podResData, nil
-}
-
-func (cpf *criFinder) Aggregate(podResData []PodResources) []v1alpha1.NUMANodeResource {
-	var perNumaRes []v1alpha1.NUMANodeResource
-
-	perNuma := make(map[int]map[v1.ResourceName]int64)
-	for nodeNum, nodeRes := range cpf.perNUMACapacity {
-		perNuma[nodeNum] = make(map[v1.ResourceName]int64)
-		for resName, resCap := range nodeRes {
-			perNuma[nodeNum][resName] = resCap
-		}
-	}
-
-	for _, podRes := range podResData {
-		for _, contRes := range podRes.Containers {
-			for _, res := range contRes.Resources {
-				cpf.updateNUMAMap(perNuma, res)
-			}
-		}
-	}
-
-	for nodeNum, resList := range perNuma {
-		numaRes := v1alpha1.NUMANodeResource{
-			NUMAID:    nodeNum,
-			Resources: make(v1.ResourceList),
-		}
-		for name, intQty := range resList {
-			numaRes.Resources[name] = *resource.NewQuantity(intQty, resource.DecimalSI)
-		}
-		perNumaRes = append(perNumaRes, numaRes)
-	}
-	return perNumaRes
 }
 
 func makeCPUResource(cpus cpuset.CPUSet) []ResourceInfo {
